src/functions/fn_index/pyproject.toml
[project]
name = "fn-index"
version = "0.1.0"
description = "Index Markdown articles into Azure AI Search"
requires-python = ">=3.11"
dependencies = [
    "azure-search-documents>=11.6.0b5",
    "azure-ai-inference>=1.0.0b1",
    "kb-ingestion-shared",
]

[tool.uv.sources]
kb-ingestion-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fn_index"]
